Clean up the code for the history pins
